Struct solana_poh::poh_recorder::PohRecorder

pub struct PohRecorder {

    pub poh: Arc<Mutex<Poh>>,
    tick_height: u64,
    clear_bank_signal: Option<Sender<bool>>,
    start_bank: Arc<Bank>,
    start_tick_height: u64,
    tick_cache: Vec<(Entry, u64)>,
    working_bank: Option<WorkingBank>,
    sender: Sender<WorkingBankEntry>,
    poh_timing_point_sender: Option<PohTimingSender>,
    leader_first_tick_height_including_grace_ticks: Option<u64>,
    leader_last_tick_height: u64,
    grace_ticks: u64,
    id: Pubkey,
    blockstore: Arc<Blockstore>,
    leader_schedule_cache: Arc<LeaderScheduleCache>,
    ticks_per_slot: u64,
    target_ns_per_tick: u64,
    record_lock_contention_us: u64,
    flush_cache_no_tick_us: u64,
    flush_cache_tick_us: u64,
    send_entry_us: u64,
    tick_lock_contention_us: u64,
    total_sleep_us: u64,
    record_us: u64,
    report_metrics_us: u64,
    ticks_from_record: u64,
    last_metric: Instant,
    record_sender: Sender<Record>,
    leader_bank_notifier: Arc<LeaderBankNotifier>,
    pub is_exited: Arc<AtomicBool>,

}

Fields
poh: Arc<Mutex<Poh>>tick_height: u64clear_bank_signal: Option<Sender<bool>>start_bank: Arc<Bank>start_tick_height: u64tick_cache: Vec<(Entry, u64)>working_bank: Option<WorkingBank>sender: Sender<WorkingBankEntry>poh_timing_point_sender: Option<PohTimingSender>leader_first_tick_height_including_grace_ticks: Option<u64>leader_last_tick_height: u64grace_ticks: u64id: Pubkeyblockstore: Arc<Blockstore>leader_schedule_cache: Arc<LeaderScheduleCache>ticks_per_slot: u64target_ns_per_tick: u64record_lock_contention_us: u64flush_cache_no_tick_us: u64flush_cache_tick_us: u64send_entry_us: u64tick_lock_contention_us: u64total_sleep_us: u64record_us: u64report_metrics_us: u64ticks_from_record: u64last_metric: Instantrecord_sender: Sender<Record>leader_bank_notifier: Arc<LeaderBankNotifier>is_exited: Arc<AtomicBool>
Implementations

impl PohRecorder

fn clear_bank(&mut self)

pub fn would_be_leader(&self, within_next_n_ticks: u64) -> bool

fn slot_for_tick_height(&self, tick_height: u64) -> Slot

pub fn leader_after_n_slots(&self, slots: u64) -> Option<Pubkey>

pub fn next_slot_leader(&self) -> Option<Pubkey>

pub fn bank(&self) -> Option<Arc<Bank>>

pub fn bank_start(&self) -> Option<BankStart>

pub fn working_bank_end_slot(&self) -> Option<Slot>

pub fn working_slot(&self) -> Option<Slot>

pub fn has_bank(&self) -> bool

pub fn tick_height(&self) -> u64

pub fn ticks_per_slot(&self) -> u64

pub fn new_recorder(&self) -> TransactionRecorder

pub fn new_leader_bank_notifier(&self) -> Arc<LeaderBankNotifier>

fn is_same_fork_as_previous_leader(&self, slot: Slot) -> bool

fn prev_slot_was_mine(&self, current_slot: Slot) -> bool

fn reached_leader_tick( &self, leader_first_tick_height_including_grace_ticks: u64 ) -> bool

pub fn start_slot(&self) -> Slot

pub fn reached_leader_slot(&self) -> PohLeaderStatus

Returns if the leader slot has been reached along with the current poh slot and the parent slot (could be a few slots ago if any previous leaders needed to be skipped).

fn compute_leader_slot_tick_heights( next_leader_slot: Option<(Slot, Slot)>, ticks_per_slot: u64 ) -> (Option<u64>, u64, u64)

fn reset_poh(&mut self, reset_bank: Arc<Bank>, reset_start_bank: bool)

pub fn reset( &mut self, reset_bank: Arc<Bank>, next_leader_slot: Option<(Slot, Slot)> )

pub fn set_bank(&mut self, bank: &Arc<Bank>, track_transaction_indexes: bool)

fn flush_cache(&mut self, tick: bool) -> Result<(), PohRecorderError>

fn report_poh_timing_point_by_tick(&self)

fn report_poh_timing_point_by_working_bank(&self, slot: Slot)

fn report_poh_timing_point(&self)

pub fn tick(&mut self)

fn report_metrics(&mut self, bank_slot: Slot)

pub fn record( &mut self, bank_slot: Slot, mixin: Hash, transactions: Vec<VersionedTransaction> ) -> Result<Option<usize>, PohRecorderError>

pub fn new_with_clear_signal( tick_height: u64, last_entry_hash: Hash, start_bank: Arc<Bank>, next_leader_slot: Option<(Slot, Slot)>, ticks_per_slot: u64, id: &Pubkey, blockstore: &Arc<Blockstore>, clear_bank_signal: Option<Sender<bool>>, leader_schedule_cache: &Arc<LeaderScheduleCache>, poh_config: &PohConfig, poh_timing_point_sender: Option<PohTimingSender>, is_exited: Arc<AtomicBool> ) -> (Self, Receiver<WorkingBankEntry>, Receiver<Record>)

pub fn new( tick_height: u64, last_entry_hash: Hash, start_bank: Arc<Bank>, next_leader_slot: Option<(Slot, Slot)>, ticks_per_slot: u64, id: &Pubkey, blockstore: &Arc<Blockstore>, leader_schedule_cache: &Arc<LeaderScheduleCache>, poh_config: &PohConfig, is_exited: Arc<AtomicBool> ) -> (Self, Receiver<WorkingBankEntry>, Receiver<Record>)

A recorder to synchronize PoH with the following data structures

bank - the LastId’s queue is updated on tick and record events

sender - the Entry channel that outputs to the ledger

pub fn get_poh_recorder_bank(&self) -> PohRecorderBank

pub fn get_working_bank_if_not_expired<'a>( bank_start: &Option<&'a BankStart> ) -> Option<&'a Arc<Bank>>

pub fn schedule_dummy_max_height_reached_failure(&mut self)